Add files via upload
This is the final project for my Python Scripting class, it is an adventure text-based game where the player navigates a dungeon to find their way out. I feel particularly proud of this project because I put a lot of effort into the systems surrounding the players inventory and how the code checks for it. I've learned a lot about dictionaries through this project as well as debugging.